Rearranged some files and formatted imports
