+++ src/disease_model/base_model.py
